api/requirements.txt
yt-dlp
google-generativeai
Flask
flask-cors
